// web/src/app/core/models/account-user-invite.model.ts
export class AccountUserInviteModel {
    AccountId?: string;
    AccountName?: string;
    CreatedAt?: string;
    CreatedBy?: string;
    DeletedAt?: string;
    Email?: string;
    InvitationCode?: string;
    IsActive?: boolean;
    RoleId?: string;
    UpdatedAt?: string;
    UpdatedBy?: string;
}