// web/src/app/core/models/account-user.model.ts
export class AccountUserModel {
    UserId?: string;
    Email?: string;
    AccountId?: string;
    RoleId?: string;
    AccountName?: string;
    RoleName?: string;
    CurrentAccount?: boolean;
    IsActive?: boolean;
    IsPending?: boolean;
    CreatedAt?: string;
    CreatedBy?: string;
    UpdatedAt?: string;
    UpdatedBy?: string;
    DeletedAt?: string;
}